Move Message Extensions to LLC
